refactor(wow): fix cross-mixin coupling and deferred imports in folder module
Move COLOR_* constants to api.py, move _call_api() to news.py where it belongs,
add comment on deferred view imports, lift or_ import to module level.
