+++ NerdyPy/modules/wow/characters.py
@@ -10,7 +10,6 @@ from blizzapi import Language, Region, RetailClient
 from discord import Color, Embed, Interaction, app_commands
 
 from modules.wow.api import (
-    RateLimited,
     check_rate_limit,
     get_best_mythic_keys,
     get_profile_link,
@@ -28,10 +27,6 @@ class WowApiLanguage(Enum):
     EN_GB = "en_GB"
 
 
-# Embed colors for guild news notifications
-COLOR_ACHIEVEMENT = Color.gold()
-COLOR_ENCOUNTER = Color.red()
-COLOR_MOUNT = Color.purple()
 COLOR_ITEM_LINK = Color(value=0x0099FF)  # WoW blue item link color
 
 
@@ -154,29 +149,6 @@ class WowCharactersMixin:
         profile_picture = "".join(asset.get("value") for asset in assets if asset.get("key") == "avatar")
 
         return character, profile_picture
-
-    async def _call_api(self, api_method, config_id, label, *args, rate_limited_event=None, stats=None, **kwargs):
-        """Call a Blizzard API method with standard rate-limit and error handling.
-
-        Returns the result on success, or None on failure (already logged).
-        Sets rate_limited_event and increments stats["skipped_error"] when provided.
-        """
-        self.bot.log.debug(f"Guild news #{config_id}: {label}")
-        try:
-            result = await asyncio.to_thread(api_method, *args, **kwargs)
-            check_rate_limit(result)
-            return result
-        except RateLimited:
-            self.bot.log.warning(f"Guild news #{config_id}: rate limited on {label}")
-            if rate_limited_event:
-                rate_limited_event.set()
-            return None
-        except Exception as ex:
-            log_fn = self.bot.log.debug if stats is not None else self.bot.log.warning
-            log_fn(f"Guild news #{config_id}: {label} failed: {ex}")
-            if stats is not None:
-                stats["skipped_error"] += 1
-            return None
 
     # ── Armory command ──────────────────────────────────────────────────
